+++ OpenSim/Actuators/Millard2012EquilibriumMuscle.h
@@ -483,11 +483,16 @@ public:
         final;
     ///@endcond
 
+    /** Adjust the properties of the muscle after the model has been scaled. The
+        optimal fiber length and tendon slack length are each multiplied by the
+        ratio of the current path length and the path length before scaling. */
+    void extendPostScale(const SimTK::State& s,
+                         const ScaleSet& scaleSet) override;
+
 //==============================================================================
 // PROTECTED METHODS
 //==============================================================================
 protected:
-    void postScale(const SimTK::State& s, const ScaleSet& aScaleSet) override;
 
     /** Gets the derivative of an actuator state by index.
         @param s The state.