+++ OpenSim/Simulation/Model/PhysicalFrame.h
@@ -135,10 +135,6 @@ public:
     // End of underlying MobilizedBody accessors.
     ///@}
 
-    /** Scale PhysicalFrame related dimensions according to predetermined 
-        ScaleFactors */
-    void scale(const SimTK::Vec3& scaleFactors);
-
     /** @name DEPRECATED API */
 
     ///@{